LPD-23983 No need of outer class, move to an inner one.
